refactor:각 도메인 안에 있는 컨트롤러 이동 - src/main/java/com/intelliJ_JO/modam/global/view
