+++ src/main/java/com/intelliJ_JO/modam/global/view/SavingsViewController.java
@@ -1,4 +1,4 @@
-package com.intelliJ_JO.modam.domain.savings.controller;
+package com.intelliJ_JO.modam.global.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
